07 fin

+++ Module07/ex01/iter.hpp
@@ -5,7 +5,7 @@
 
 template<typename T>
 void iter(T* arr, size_t arr_len, void (*f)(T)) {
-	for(int i = 0; i < arr_len; i++)
+	for(size_t i = 0; i < arr_len; i++)
 		(*f)(arr[i]);
 }
